requirements.txt
pillow
streamlit
anthropic
faster-whisper
moviepy
opencv-python
numpy
mcp